moved handlers to handlers.go and hide password functions from export
